refactor: extracted react query hooks  and moved into into /hooks/api/
